Webapp supports subtitle track selector
