+++ common/app/components/LandingPage.tsx
@@ -2,23 +2,27 @@ import React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { makeStyles } from '@mui/styles';
 import gt from 'semver/functions/gt';
+import Box from '@mui/material/Box';
 import Fade from '@mui/material/Fade';
 import Paper from '@mui/material/Paper';
 import Link from '@mui/material/Link';
 import Typography from '@mui/material/Typography';
 import Button from '@mui/material/Button';
 import ChromeExtension from '../services/chrome-extension';
-import { useMediaQuery, useTheme, type Theme } from '@mui/material';
+import { useTheme, type Theme } from '@mui/material/styles';
+import useMediaQuery from '@mui/material/useMediaQuery';
 import { useAppBarHeight } from '../../hooks/use-app-bar-height';
 import { VideoTabModel } from '../..';
 import VideoElementSelector from './VideoElementSelector';
+import LoadSubtitlesIcon from '../../components/LoadSubtitlesIcon';
+import RestoreIcon from '@mui/icons-material/Restore';
 
 interface StylesProps {
     appBarHidden: boolean;
     appBarHeight: number;
 }
 
-const useStyles = makeStyles<Theme, StylesProps>((theme) => ({
+const useStyles = makeStyles<Theme, StylesProps>({
     background: ({ appBarHidden, appBarHeight }) => ({
         position: 'absolute',
         height: appBarHidden ? '100vh' : `calc(100vh - ${appBarHeight}px)`,
@@ -34,17 +38,7 @@ const useStyles = makeStyles<Theme, StylesProps>((theme) => ({
     browseLink: {
         cursor: 'pointer',
     },
-    actionsContainer: {
-        position: 'absolute',
-        bottom: 0,
-        left: 0,
-        padding: theme.spacing(2),
-        width: '100%',
-        display: 'flex',
-        flexDirection: 'column',
-        gap: theme.spacing(1),
-    },
-}));
+});
 
 interface Props {
     extension: ChromeExtension;
@@ -60,6 +54,7 @@ interface Props {
         React.MouseEventHandler<HTMLLabelElement>;
     onVideoElementSelected: (videoElement: VideoTabModel) => void;
     onRestoreLastSession: () => void;
+    onOpenSubtitleTrackSelector: () => void;
 }
 
 export default function LandingPage({
@@ -74,6 +69,7 @@ export default function LandingPage({
     onFileSelector,
     onVideoElementSelected,
     onRestoreLastSession,
+    onOpenSubtitleTrackSelector,
 }: Props) {
     const { t } = useTranslation();
     const appBarHeight = useAppBarHeight();
@@ -81,6 +77,15 @@ export default function LandingPage({
     const extensionUpdateAvailable = extension.version && gt(latestExtensionVersion, extension.version);
     const theme = useTheme();
     const smallScreen = useMediaQuery(theme.breakpoints.down(500));
+    const showVideoElementSelector =
+        extension.supportsLandingPageStreamingVideoElementSelector && videoElements.length > 0;
+    let buttonCount = 1;
+    if (canRestoreLastSession) {
+        buttonCount++;
+    }
+    if (showVideoElementSelector) {
+        buttonCount++;
+    }
 
     return (
         <Paper square className={classes.background}>
@@ -121,19 +126,53 @@ export default function LandingPage({
                     </Typography>
                     {(canRestoreLastSession ||
                         (extension.supportsLandingPageStreamingVideoElementSelector && videoElements.length > 0)) && (
-                        <div className={classes.actionsContainer}>
-                            {canRestoreLastSession && (
-                                <Button variant="outlined" color="primary" onClick={onRestoreLastSession} fullWidth>
-                                    {t('landing.restoreLastSession')}
+                        <Box
+                            sx={{
+                                position: 'absolute',
+                                display: 'flex',
+                                flexDirection: 'column',
+                                bottom: 0,
+                                left: 0,
+                                padding: 1.5,
+                                width: '100%',
+                                gap: 1.5,
+                            }}
+                        >
+                            <Box
+                                sx={{
+                                    display: 'grid',
+                                    gridTemplateColumns: { xs: '1fr', md: `repeat(${buttonCount}, 1fr)` },
+                                    gap: 1.5,
+                                }}
+                            >
+                                <Button
+                                    variant="outlined"
+                                    color="primary"
+                                    startIcon={<LoadSubtitlesIcon fontSize="small" />}
+                                    onClick={onOpenSubtitleTrackSelector}
+                                    fullWidth
+                                >
+                                    {t('action.loadSubtitles')}
                                 </Button>
-                            )}
-                            {extension.supportsLandingPageStreamingVideoElementSelector && videoElements.length > 0 && (
-                                <VideoElementSelector
-                                    videoElements={videoElements}
-                                    onVideoElementSelected={onVideoElementSelected}
-                                />
-                            )}
-                        </div>
+                                {canRestoreLastSession && (
+                                    <Button
+                                        variant="outlined"
+                                        color="primary"
+                                        startIcon={<RestoreIcon />}
+                                        onClick={onRestoreLastSession}
+                                        fullWidth
+                                    >
+                                        {t('landing.restoreLastSession')}
+                                    </Button>
+                                )}
+                                {showVideoElementSelector && videoElements.length > 0 && (
+                                    <VideoElementSelector
+                                        videoElements={videoElements}
+                                        onVideoElementSelected={onVideoElementSelected}
+                                    />
+                                )}
+                            </Box>
+                        </Box>
                     )}
                 </div>
             </Fade>
