+++ common/components/OnlineSubtitleSourceDialog.tsx
@@ -19,7 +19,7 @@ import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { JimakuClient, JimakuEntry } from '@/services/subtitle-sources';
+import { JimakuClient, JimakuEntry } from '@project/common/subtitle-sources';
 import type { JimakuCachedWork } from '@project/common/global-state';
 import IconButton from '@mui/material/IconButton';
 import InputAdornment from '@mui/material/InputAdornment';
