+++ extension/src/controllers/video-data-sync-controller.ts
@@ -256,7 +256,6 @@ export default class VideoDataSyncController {
                   suggestedName: document.title,
                   selectedSubtitle: autoSelectedTrackIds,
                   error: '',
-                  showSubSelect: true,
                   subtitles: subtitleTrackChoices,
                   defaultCheckboxState: defaultCheckboxState,
                   openedFromAsbplayerId: '',
@@ -533,13 +532,13 @@ export default class VideoDataSyncController {
             const subtitles: SerializedSubtitleFile[] = [];
 
             for (let i = 0; i < data.length; i++) {
-                const { extension, url, language, localFile } = data[i];
+                const { extension, url, language, file } = data[i];
                 const subtitleFiles = await this._subtitlesForUrl(
                     this._defaultVideoName(this._syncedData?.basename, data[i]),
                     language,
                     extension,
-                    url,
-                    localFile
+                    url!,
+                    file !== undefined
                 );
                 if (subtitleFiles !== undefined) {
                     subtitles.push(...subtitleFiles);
@@ -565,8 +564,8 @@ export default class VideoDataSyncController {
             const subtitles: SerializedSubtitleFile[] = [];
 
             for (let i = 0; i < data.length; i++) {
-                const { name, language, extension, url, localFile } = data[i];
-                const subtitleFiles = await this._subtitlesForUrl(name, language, extension, url, localFile);
+                const { name, language, extension, url, file } = data[i];
+                const subtitleFiles = await this._subtitlesForUrl(name, language, extension, url!, file !== undefined);
                 if (subtitleFiles !== undefined) {
                     subtitles.push(...subtitleFiles);
                 }
@@ -702,7 +701,6 @@ export default class VideoDataSyncController {
         return client.updateState({
             open: true,
             isLoading: false,
-            showSubSelect: true,
             error,
             themeType: themeType,
         });
